Added cpp folder for arijit.

+++ storage_engine/include/TransitiveClosure.hpp
@@ -55,7 +55,7 @@ namespace tc {
 
     uint64_t * set_data = ((uint64_t*)vs->get_data())+1;
     const TrieBlock<T,M> * input_head = input->getHead();
-    size_t iteration = 0;
+    int iteration = 0;
     while(frontier_size != 0 && (iterations < 0 || iteration < iterations) ){
       std::cout  << "ITERATION: " << iteration << " FRONTIER SIZE: " << frontier_size << std::endl;
       par::for_range(0,frontier_size,100,[&](const size_t tid, const size_t i){
